Move Controls to a separate file.
